backend/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
sqlalchemy>=2.0
psycopg[binary,pool]>=3.2
pydantic>=2.7
python-multipart>=0.0.9
passlib[bcrypt]>=1.7
pyjwt[crypto]>=2.9
email-validator>=2.1
python-dateutil>=2.9
orjson>=3.10
supabase